backend/requirements.txt
flask>=3.0.0
flask-socketio>=5.3.6
flask-cors>=4.0.0
pymongo>=4.6.0
marshmallow>=3.20.0
firebase-admin>=6.3.0
python-dotenv>=1.0.0
flask-talisman>=1.1.0
pytest>=8.0.0
mongomock>=4.1.0
pillow>=10.0.0
python-magic>=0.4.27
flask-limiter>=3.5.0
pytest-cov>=4.1.0